barbershop_project/pyproject.toml
[tool.poetry]
name = "barbershop-project"
version = "0.1.0"
description = ""
authors = ["Rinat <tensal@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"